// src/types/conference.ts
export type PresentationType = "ORAL" | "POSTER" | "WORKSHOP" | "OTHER";

export interface AbstractSubmissionRequest {
  firstName: string;
  lastName: string;
  email: string;
  phone?: string | undefined;
  organization?: string | undefined;
  country?: string | undefined;
  abstractTitle: string;
  abstractBody: string;
  keywords?: string | undefined;
  presentationType: PresentationType;
}

export interface AbstractSubmissionResponse {
  abstractNumber: string;
  submissionStatus: string;
  email: string;
}